demo/reset-demo-function/requirements.txt
cachetools==4.0.0
certifi==2020.4.5
chardet==3.0.4
google-api-core==1.16.0
google-auth==1.13.1
google-cloud-container==0.4.0
googleapis-common-protos==1.51.0
grpc-google-iam-v1==0.12.3
grpcio==1.28.1
idna==2.9
kubernetes==11.0.0
oauthlib==3.1.0
protobuf==3.11.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
python-dateutil==2.8.1
pytz==2019.3
pyyaml==5.3.1
requests==2.23.0
requests-oauthlib==1.3.0
rsa==4.0
six==1.14.0
urllib3==1.25.8
websocket-client==0.57.0
